requirements.txt
# Add Python dependencies here
yt-dlp
pandas
shutil
whisper
torch
cerebras-cloud-sdk
tqdm
python-dotenv
pyyaml